+++ workspace_1.16.1/STM32F405_EPS/Debug/objects.list
@@ -3,6 +3,7 @@
 "./Core/Src/dma.o"
 "./Core/Src/gpio.o"
 "./Core/Src/i2c.o"
+"./Core/Src/ina219.o"
 "./Core/Src/main.o"
 "./Core/Src/mppt.o"
 "./Core/Src/pdu.o"
